magit--insert-modules-overview: don't try to be smart about version strings
This solution has seemed to only care about the "vNUM" case, but the
world of version strings is much more colourful. Tags often start with
more letters (or words such as "release" or "debian"), not just one
followed by numbers and punctuation, so the alignment wasn't perfect
anyway.

Rather than trying to handle one special case, align everything to the
same column and be done with it.