magit--insert-modules-overview: try to optimize column alignment
Compute the path column width according to the longest path, with a
cap at half the window width.

The assumption is that while tags and branch names rarely exceed 20
characters and typically fit into 10, module paths around 30 chars are
not so uncommon.

I can imagine how subjective this might be or other possible
optimization approaches, but at least the exact fitting of the path
column seems a definite improvement (and a simple one, as the modules
list is already there).